requirements.txt
django-crispy-forms
crispy-bootstrap5
django-debug-toolbar
django-filter
whitenoise
django-allauth
celery
django-extensions
psycopg[binary]
python-dotenv
django-tables2
django-htmx
djangorestframework==3.14.0